+++ tests/Common/AbstractSQLDumpFileTest.php
@@ -8,11 +8,10 @@ use PHPUnit\Framework\TestCase;
 use Throwable;
 use Yiisoft\Cache\Db\DbCache;
 use Yiisoft\Db\Connection\ConnectionInterface;
-use Yiisoft\Db\Constraint\IndexConstraint;
+use Yiisoft\Db\Constant\ColumnType;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidConfigException;
 use Yiisoft\Db\Exception\NotSupportedException;
-use Yiisoft\Db\Schema\SchemaInterface;
 
 /**
  * @group Mssql
@@ -71,12 +70,12 @@ abstract class AbstractSQLDumpFileTest extends TestCase
 
         $schema = $this->db->getSchema();
 
-        /** @psalm-var IndexConstraint[] $indexes */
-        $indexes = $schema->getTableIndexes($dbCache->getTable(), true);
+        $index = $schema->getTablePrimaryKey($dbCache->getTable(), true);
 
-        $this->assertSame(['id'], $indexes[0]->getColumnNames());
-        $this->assertTrue($indexes[0]->isUnique());
-        $this->assertTrue($indexes[0]->isPrimary());
+        $this->assertNotNull($index);
+        $this->assertSame(['id'], $index->columnNames);
+        $this->assertTrue($index->isPrimaryKey);
+        $this->assertTrue($index->isUnique);
 
         $this->loadFromSQLDumpFile(dirname(__DIR__, 2) . "/sql/$this->driverName-down.sql");
 
@@ -96,17 +95,17 @@ abstract class AbstractSQLDumpFileTest extends TestCase
         $this->loadFromSQLDumpFile(dirname(__DIR__, 2) . "/sql/$this->driverName-up.sql");
 
         $tableSchema = $this->db->getTableSchema($dbCache->getTable());
-        $tableRawName = $this->db->getSchema()->getRawTableName($dbCache->getTable());
+        $tableRawName = $this->db->getQuoter()->getRawTableName($dbCache->getTable());
 
         $this->assertNotNull($tableSchema);
 
         $this->assertSame($tableRawName, $tableSchema->getName());
         $this->assertSame(['id'], $tableSchema->getPrimaryKey());
         $this->assertSame(['id', 'data', 'expire'], $tableSchema->getColumnNames());
-        $this->assertSame(SchemaInterface::TYPE_STRING, $tableSchema->getColumn('id')?->getType());
+        $this->assertSame(ColumnType::STRING, $tableSchema->getColumn('id')?->getType());
         $this->assertSame(128, $tableSchema->getColumn('id')?->getSize());
-        $this->assertSame(SchemaInterface::TYPE_BINARY, $tableSchema->getColumn('data')?->getType());
-        $this->assertSame(SchemaInterface::TYPE_INTEGER, $tableSchema->getColumn('expire')?->getType());
+        $this->assertSame(ColumnType::BINARY, $tableSchema->getColumn('data')?->getType());
+        $this->assertSame(ColumnType::INTEGER, $tableSchema->getColumn('expire')?->getType());
 
         $this->loadFromSQLDumpFile(dirname(__DIR__, 2) . "/sql/$this->driverName-down.sql");