+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.0.1 under development
 
+- Chg #88: Bump supported version of `yiisoft/db` to `^2.0` (@batyrmastyr)
 - Chg #91: Change supported PHP versions to `8.1 - 8.4` (@batyrmastyr)
 
 1.0.0 May 09, 2023